Refactor syndicate/affinity/rent into focused engine modules
Extract rent resolution, post-move phase helpers, syndicate formation,
and affinity action handlers from gameStateMachine to reduce file size
and duplicate control flow. Wire disruption nullify resume through
resolvePostMovePhase.
